script/requirements.txt
requests
google-cloud-bigquery
flask
gunicorn